+++ zerver/migrations/0001_squashed_0569.py
@@ -22,6 +22,12 @@ import zerver.models.realms
 import zerver.models.streams
 
 
+class LegacyBotCreationPolicyEnum(IntEnum):
+    EVERYONE = 1
+    LIMIT_GENERIC_BOTS = 2
+    ADMINS_ONLY = 3
+
+
 class LegacyCommonMessagePolicyEnum(IntEnum):
     MEMBERS_ONLY = 1
     ADMINS_ONLY = 2
@@ -831,7 +837,7 @@ class Migration(migrations.Migration):
                 (
                     "bot_creation_policy",
                     models.PositiveSmallIntegerField(
-                        default=zerver.models.realms.BotCreationPolicyEnum["EVERYONE"]
+                        default=LegacyBotCreationPolicyEnum["EVERYONE"]
                     ),
                 ),
                 ("custom_upload_quota_gb", models.IntegerField(null=True)),